Core.UI.WPF, UI, UI.TimeTracking - fix hot key issues and allow deletion of CategoryAssignmentGrid
There was an issue which caused an NullReferenceException while trying to register the hot key. This issue has been resolved with a null check before executing the register or unregister of an hot key.

It should be possible to delete Categories now. There is a new menu item in the context menu for categories which should provide this functionality.